+++ src/main/java/com/cauchynote/friend/mapper/FriendMapper.java
@@ -21,7 +21,7 @@ public interface FriendMapper {
      * @return 好友ID列表
      */
     @Select("select friend_ids from note_friend where user_id = #{userId}")
-    String getFriendIds(Long userId);
+    String getFriendIds(Integer userId);
 
     /**
      * 获取好友列表
@@ -39,7 +39,7 @@ public interface FriendMapper {
         "</foreach> limit #{startNum}, #{pageSize} " +
         "</script>")
     @Results(id = "getFriendList", value = {
-        @Result(column = "id", property = "id", javaType = Long.class, jdbcType = JdbcType.BIGINT),
+        @Result(column = "id", property = "id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
         @Result(column = "user_name", property = "username", javaType = String.class, jdbcType = JdbcType.VARCHAR),
         @Result(column = "email", property = "email", javaType = String.class, jdbcType = JdbcType.VARCHAR),
         @Result(column = "password", property = "password", javaType = String.class, jdbcType = JdbcType.VARCHAR),
@@ -49,7 +49,7 @@ public interface FriendMapper {
         @Result(column = "is_password_non_expired", property = "isPasswordNonExpired", javaType = Integer.class, jdbcType = JdbcType.TINYINT),
         @Result(column = "is_enable", property = "isEnable", javaType = Integer.class, jdbcType = JdbcType.TINYINT)
     })
-    List<User> getFriendList(@Param("list") List<Long> friendIds, Integer startNum, Integer pageSize);
+    List<User> getFriendList(@Param("list") List<Integer> friendIds, Integer startNum, Integer pageSize);
 
     /**
      * 更新好友信息
@@ -59,7 +59,7 @@ public interface FriendMapper {
      * @return 1- 成功 0 - 失败
      */
     @Update("update note_friend set friend_ids = #{friendIds} where user_id = #{userId}")
-    Integer updateFriend(Long userId, String friendIds);
+    Integer updateFriend(Integer userId, String friendIds);
 
     /**
      * 查询好友
@@ -70,7 +70,7 @@ public interface FriendMapper {
     @Select("select id, user_name, email, password, create_time, is_non_expired, is_non_locked, is_password_non_expired, " +
         "is_enable from note_user where user_name = #{friendName}")
     @Results(id = "searchFriendByUsername", value = {
-        @Result(column = "id", property = "id", javaType = Long.class, jdbcType = JdbcType.BIGINT),
+        @Result(column = "id", property = "id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
         @Result(column = "user_name", property = "username", javaType = String.class, jdbcType = JdbcType.VARCHAR),
         @Result(column = "email", property = "email", javaType = String.class, jdbcType = JdbcType.VARCHAR),
         @Result(column = "password", property = "password", javaType = String.class, jdbcType = JdbcType.VARCHAR),
@@ -89,7 +89,7 @@ public interface FriendMapper {
      * @param friendIds 好友ID
      */
     @Insert("insert into note_friend(user_id, friend_ids) value(#{userId}, #{friendIds})")
-    Integer addNewRecord(Long userId, String friendIds);
+    Integer addNewRecord(Integer userId, String friendIds);
 
     /**
      * 新增好友请求记录
@@ -98,7 +98,7 @@ public interface FriendMapper {
      * @param friendRequestIds 好友ID
      */
     @Insert("insert into note_friend(user_id, friend_request_ids) value(#{userId}, #{friendRequestIds})")
-    void addNewFriendRequestRecord(Long userId, String friendRequestIds);
+    void addNewFriendRequestRecord(Integer userId, String friendRequestIds);
 
     /**
      * 获取好友请求ID列表
@@ -107,7 +107,7 @@ public interface FriendMapper {
      * @return 好友请求ID列表
      */
     @Select("select friend_request_ids from note_friend where user_id = #{userId}")
-    String getFriendRequestIds(Long userId);
+    String getFriendRequestIds(Integer userId);
 
     /**
      * 更新好友请求信息
@@ -117,7 +117,7 @@ public interface FriendMapper {
      * @return 1- 成功 0 - 失败
      */
     @Update("update note_friend set friend_request_ids = #{friendRequestIds} where user_id = #{userId}")
-    Integer updateFriendRequest(Long userId, String friendRequestIds);
+    Integer updateFriendRequest(Integer userId, String friendRequestIds);
 
     /**
      * 查询好友请求列表
@@ -133,7 +133,7 @@ public interface FriendMapper {
         "</foreach> " +
         "</script>")
     @Results(id = "getFriendRequestList", value = {
-        @Result(column = "id", property = "id", javaType = Long.class, jdbcType = JdbcType.BIGINT),
+        @Result(column = "id", property = "id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
         @Result(column = "user_name", property = "username", javaType = String.class, jdbcType = JdbcType.VARCHAR),
         @Result(column = "email", property = "email", javaType = String.class, jdbcType = JdbcType.VARCHAR),
         @Result(column = "password", property = "password", javaType = String.class, jdbcType = JdbcType.VARCHAR),
@@ -143,7 +143,7 @@ public interface FriendMapper {
         @Result(column = "is_password_non_expired", property = "isPasswordNonExpired", javaType = Integer.class, jdbcType = JdbcType.TINYINT),
         @Result(column = "is_enable", property = "isEnable", javaType = Integer.class, jdbcType = JdbcType.TINYINT)
     })
-    List<User> getFriendRequestList(@Param("list") List<Long> friendRequestIds);
+    List<User> getFriendRequestList(@Param("list") List<Integer> friendRequestIds);
 
     /**
      * 查询记录数量
@@ -152,5 +152,5 @@ public interface FriendMapper {
      * @return 记录数量
      */
     @Select("select count(*) from note_friend where user_id = #{userId}")
-    Integer getRecordCountOfUser(Long userId);
+    Integer getRecordCountOfUser(Integer userId);
 }
